Async exec handler.
